requirements.txt
streamlit
pandas
numpy
scikit-learn
matplotlib
seaborn
plotly
langchain
langchain-experimental
langchain-google-genai
tabulate
xgboost
shap
statsmodels
fpdf2
